requirements.txt
####### Required Libraries for the Application #######
mysql-connector-python
docx
tk